+++ bambi/tests/test_built_models.py
@@ -1,9 +1,11 @@
+import pytest
+
 import arviz as az
 import numpy as np
 import pandas as pd
-import pytest
 import theano.tensor as tt
 
+from bambi import math
 from bambi.models import Model
 from bambi.priors import Prior
 
@@ -519,3 +521,34 @@ def test_model_graph(crossed_data):
         model.graph()
     model.build()
     model.graph()
+
+
+def test_potentials():
+    data = pd.DataFrame(np.repeat((0, 1), (18, 20)), columns=["w"])
+    priors = {"Intercept": Prior("Uniform", lower=0, upper=1)}
+    potentials = [
+        (("Intercept", "Intercept"), lambda x, y: math.switch(x < 0.45, y, -np.inf)),
+        ("Intercept", lambda x: math.switch(x > 0.55, 0, -np.inf)),
+    ]
+
+    model = Model(
+        "w ~ 1",
+        data,
+        family="bernoulli",
+        link="identity",
+        priors=priors,
+        potentials=potentials,
+    )
+    model.build()
+    assert len(model.backend.model.potentials) == 2
+
+    pot0 = model.backend.model.potentials[0].get_parents()[0]
+    pot1 = model.backend.model.potentials[1].get_parents()[0]
+    pot0.__str__() == (
+        "Elemwise{switch,no_inplace}(Elemwise{lt,no_inplace}.0, "
+        "Intercept ~ Uniform, TensorConstant{-inf})"
+    )
+    pot1.__str__() == (
+        "Elemwise{switch,no_inplace}(Elemwise{gt,no_inplace}.0, "
+        "TensorConstant{0}, TensorConstant{-inf})"
+    )